Break up the "JSON upstream" test suite into multiple modules
One module per type is good.
